SingleRequest refactoring
- move functions into request.[ch]
- differentiate between reset and free
- add Curl_req_done() to perform last actions
- add a send bufq to SingleRequest for future
  use in keeping upload data
